Fix more warings (#3396)

+++ native/cocos/platform/win32/Device-win32.cpp
@@ -40,7 +40,7 @@ int Device::getDPI() {
         int PixelsX = GetDeviceCaps(hScreenDC, HORZRES);
         int MMX = GetDeviceCaps(hScreenDC, HORZSIZE);
         ReleaseDC(nullptr, hScreenDC);
-        dpi = 254.0f * PixelsX / MMX / 10;
+        dpi = static_cast<int>(254.0f * PixelsX / MMX / 10);
     }
     return dpi;
 }
